Some more moves and refactors
